renamed non-standard 'lib' to '_lib'
